backend/pyproject.toml
[project]
name = "ai-accountant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.6",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
]
